+++ app/components/table_components/table.rb
@@ -4,19 +4,21 @@ class TableComponents::Table < ViewComponent::Base
 
   erb_template <<~ERB
     <div class="flex bg-white justify-between items-center px-4 border-b border-gray-200">
-        <div class="flex-1">
-    <% if left? %>
-        <%= left %>
-    <% end %>
-        </div>
+      <div class="flex-1">
+        <% if left? %>
+          <%= left %>
+        <% end %>
+      </div>
       <%= render CommonComponents::PaginationComponent.new(pagy: @pagy) unless @options[:no_pagination] || @pagy.nil? %>
       </div>
       <div class="overflow-x-scroll bg-white">
-        <div class="grid" style="<%= grid_columns %>">
+        <div id="<%= @options[:turbo_id] %>" class="grid" style="<%= grid_columns %>">
           <%= render TableComponents::Column.with_collection(@row_component::columns(**@column_arguments), order_scope_name: @order_scope_name) %>
+          <% if @options[:turbo_id] %> <div id="turbo-separator" class='h-px' style="grid-column: 1/-1;"></div> <% end %>
           <%= render @row_component.with_collection(@rows, pagy: @pagy, **@row_arguments) %>
         </div>
       </div>
+    </div>
   ERB
 
   # rubocop:disable Metrics/ParameterLists